Reformat DisputeModelTest
